% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdp.R
\name{cdp_run}
\alias{cdp_run}
\title{Run ChemiDPlus on all HemOnc Drugs}
\usage{
cdp_run(
  conn,
  conn_fun,
  vocab_schema,
  steps = c("log_registry_number", "get_rn_url_validity", "get_classification",
    "get_names_and_synonyms", "get_registry_numbers", "get_links_to_resources"),
  search_type = "contains",
  sleep_time = 5,
  verbose = TRUE,
  render_sql = TRUE
)
}
\description{
Unlike \code{\link{cg_run}}, which runs the on the entire NCI Drug Dictionary, the constraints on the ChemiDPlus feature is the names of all HemOnc Drugs.
}
