% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-cdp-omop.R
\name{cdp_to_omop}
\alias{cdp_to_omop}
\title{ETL the ChemiDPlus Tables to OMOP Vocabulary Architecture}
\usage{
cdp_to_omop(
  conn,
  destination_schema,
  vocab_schema,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{destination_schema}{Schema where the new tables will be written to. User must have write privileges for this schema}

\item{vocab_schema}{Schema in the connection that houses the OMOP Proper Vocabularies. Cannot be the same as the \code{destination_schema} because the OMOP Proper Vocabularies would be appended with the ChemiDPlus transformations and it needs to be silo'd until the concepts are vetted.}
}
\description{
ETL the ChemiDPlus Tables to OMOP Vocabulary Architecture
}
\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{readTable}},\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[rubix]{map_names_set}}, \code{\link[rubix]{normalize_all_to_na}},
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{group_by}}
\code{\link[tidyr]{extract}}
\code{\link[chariot]{queryAthena}},\code{\link[chariot]{filterValid}}
\code{\link[stringr]{str_remove}}
}
