% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg.R
\name{cg_search}
\alias{cg_search}
\title{Search Cancer.gov}
\usage{
cg_search(search_term, size = 1000, matchType = "Begins", crawl_delay = 5)
}
\arguments{
\item{search_term}{PARAM_DESCRIPTION}

\item{size}{PARAM_DESCRIPTION, Default: 1000}

\item{matchType}{PARAM_DESCRIPTION, Default: 'Begins'}

\item{crawl_delay}{PARAM_DESCRIPTION, Default: 5}
}
\description{
Search the NCI Drug Dictionary using the API. Note that the results are not stored in the database in this version in the manner that ChemiDPlus (\code{cdp_search}) does in this package version.
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
