% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_tables_to_omop.R
\name{chemidplus_tables_to_omop}
\alias{chemidplus_tables_to_omop}
\title{ETL the ChemiDPlus Tables to OMOP Vocabulary Architecture}
\usage{
chemidplus_tables_to_omop(
  conn,
  file_report = TRUE,
  file_report_to = paste0("chemidplus_tables_to_omop_", Sys.Date(), ".txt"),
  append_file_report = TRUE
)
}
\description{
ETL the ChemiDPlus Tables to OMOP Vocabulary Architecture
}
\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{readTable}},\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[rubix]{map_names_set}},\code{\link[rubix]{mutate_all_as_char}},\code{\link[rubix]{normalize_all_to_na}},\code{\link[rubix]{make_identifier}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{group_by}}
\code{\link[tidyr]{extract}}
\code{\link[chariot]{queryAthena}},\code{\link[chariot]{filterValid}}
\code{\link[stringr]{str_remove}}
}
