% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicaltrialsgov_functions.R
\name{clinicaltrialsgov_functions}
\alias{clinicaltrialsgov_functions}
\title{ClinicalTrials.gov Functions}
\description{
The ClinicalTrials.gov functions are related to the `aact` Postgres database instance available for download at \href{https://aact.ctti-clinicaltrials.org/}{https://aact.ctti-clinicaltrials.org/}.
}
\section{AACT Database}{

Technical documentation on the AACT Database instantiation can be found at \href{https://aact.ctti-clinicaltrials.org/snapshots}{https://aact.ctti-clinicaltrials.org/snapshots}. Documentation on database structure and naming conventions can be found at \href{https://aact.ctti-clinicaltrials.org/schema}{https://aact.ctti-clinicaltrials.org/schema}.
}

