% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCancerGovSchema.R
\name{createCancerGovSchema}
\alias{createCancerGovSchema}
\title{Create CancerGov Schema}
\usage{
createCancerGovSchema(conn)
}
\description{
This function creates the "cancergov" schema if it doesn't exist
}
