% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDefinitionTable.R
\name{createDefinitionTable}
\alias{createDefinitionTable}
\title{Create CONCEPT_DEFINITION Table}
\usage{
createDefinitionTable(conn = conn, .input)
}
\description{
The CONCEPT_DEFINITION Table is a custom addition to the OMOP Vocabulary Infrastructure and stores the human readable Data Dictionary definitions found in Cancer.gov's site.
}
