% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetadataTable.R
\name{createMetadataTable}
\alias{createMetadataTable}
\title{Create CONCEPT_METADATA Table}
\usage{
createMetadataTable(conn, .input)
}
\description{
The CONCEPT_METADATA Table is a custom addition to the OMOP Vocabulary Infrastructure and stores the metadata associated with the scrape of a given concept.
}
