% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddlCancerGovSchema.R
\name{ddlCancerGovSchema}
\alias{ddlCancerGovSchema}
\title{DDL Cancer Gov Schema}
\usage{
ddlCancerGovSchema(conn)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
If the cancergov schema does not have any tables, writes empty OMOP Vocabulary Tables
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{send}}
 \code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
}
