% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_schema_to_data_repo.R
\name{export_schema_to_data_repo}
\alias{export_schema_to_data_repo}
\title{Export a Schema in a Postgres Database to a Repo}
\usage{
export_schema_to_data_repo(target_dir, schema)
}
\arguments{
\item{target_dir}{Local repo where the schema data will be exported to. If it does not already exist on the local drive, it should be cloned beforehand.}

\item{schema}{Schema}
}
\description{
There should be a Data Repo on the local machine that is linked to the schema.
}
\seealso{
\code{\link[cave]{create_dir_if_not_exist}}
 \code{\link[chariot]{connectAthena}},\code{\link[chariot]{dcAthena}}
 \code{\link[pg13]{lsTables}},\code{\link[pg13]{readTable}}
 \code{\link[rubix]{map_names_set}}
 \code{\link[purrr]{map2}},\code{\link[purrr]{map}}
 \code{\link[broca]{simply_write_csv}}
 \code{\link[readr]{read_lines}}
 \code{\link[sinew]{makeOxygen}}
 \code{\link[stringr]{str_replace}}
 \code{\link[glitter]{docPushInstall}}
}
