% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_functions.R
\name{get_classification_code}
\alias{get_classification_code}
\title{Scrape the Classification Code in the Summary Header of the RN URL}
\usage{
get_classification_code(
  conn,
  rn_url,
  response,
  schema = "chemidplus",
  sleep_time = 3
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{rn_url}{PARAM_DESCRIPTION}

\item{sleep_time}{PARAM_DESCRIPTION, Default: 3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xml2]{read_xml}}
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}}
\code{\link[stringr]{str_remove}}
\code{\link[centipede]{strsplit}}
\code{\link[purrr]{map2}},\code{\link[purrr]{set_names}},\code{\link[purrr]{map}}
}
