% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R
\name{get_dictionary_and_links}
\alias{get_dictionary_and_links}
\title{Scrape the Drug Definitions and Links from the NCI Drug Dictionary}
\usage{
get_dictionary_and_links(
  conn,
  max_page = 50,
  sleep_time = 3,
  verbose = TRUE,
  render_sql = TRUE,
  crawl_delay = 5,
  size = 10000
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{max_page}{maximum page number to iterate the scrape over in the "https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=" path, Default: 50}

\item{sleep_time}{Time in seconds for the system to sleep before each scrape with \code{\link[xml2]{read_html}}.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\value{
Any differences found between the scraped data and the existing data in the Drug Dictionary and Drug Link Tables are appended to their respective tables with the local timestamp.
}
\description{
Run the full sequence that scrapes, parses, and stores the NCI Drug Dictionary found at CancerGov.org and any correlates to the NCI Thesaurus in a Postgres Database.
}
\details{
Scrapes the Definitions and the links to each Drug Page at the main Drug Dictionary pages in  the \url{https://www.cancer.gov/publications/dictionaries/cancer-drug}\emph{{i}} and stores the parsed response to the Drug Dictionary and Drug Link Tables, respectively.
}
\section{Web Source Types}{

The NCI Drug Dictionary has 2 data sources that run in parallel. The first source is the Drug Dictionary itself at \href{https://www.cancer.gov/publications/dictionaries/cancer-drug}{https://www.cancer.gov/publications/dictionaries/cancer-drug}. The other source are the individual drug pages, called Drug Detail Links in skyscraper, that contain tables of synonyms, including investigational names.
}

\section{Drug Dictionary}{

The listed drug names and their definitions are scraped from the Drug Dictionary HTML and updated to a Drug Dictionary Table in a \code{cancergov} schema.
}

\seealso{
\code{\link[pg13]{brake_closed_conn}},\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[secretary]{typewrite_progress}},\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{character(0)}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{distinct}}
}
