% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drug.R
\name{get_drug}
\alias{get_drug}
\title{Scrape the NCI Drug Dictionary}
\usage{
get_drug(max_page = 50)
}
\arguments{
\item{max_page}{Maximum page to scrape, Default: 39}
}
\value{
A data frame
}
\description{
Scrape the drug names and their definitions from all the pages of the NCI Drug Dictionary (https://www.cancer.gov/publications/dictionaries/cancer-drug)
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[secretary]{typewrite}}
 \code{\link[xml2]{read_xml}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
 \code{\link[purrr]{map}},\code{\link[purrr]{keep}},\code{\link[purrr]{transpose}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}}
 \code{\link[stringr]{str_replace}}
}
