% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drug_link_synonyms.R
\name{get_drug_link_synonym}
\alias{get_drug_link_synonym}
\title{Scrape a Drug HTML Page}
\usage{
get_drug_link_synonym(conn, sleep_time = 5)
}
\description{
Scrape the Tables in the Drug Page of the NCI Drug Dictionary if one exists
}
\seealso{
\code{\link[dplyr]{select_all}}
 \code{\link[tibble]{rownames}},\code{\link[tibble]{c("tibble", "tibble")}}
 \code{\link[progress]{progress_bar}}
 \code{\link[secretary]{typewrite}},\code{\link[secretary]{character(0)}}
 \code{\link[xml2]{read_xml}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
}
