% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_and_synonyms.R
\name{get_names_and_synonyms}
\alias{get_names_and_synonyms}
\title{FUNCTION_TITLE}
\usage{
get_names_and_synonyms(
  conn,
  rn_url,
  response,
  schema = "chemidplus",
  sleep_time = 3
)
}
\arguments{
\item{conn}{Postgres connection object}

\item{rn_url}{Registry number URL to read}

\item{response}{(optional) "xml_document" "xml_node" class object returned by xml2::read_html for the `rn_url` argument. Providing a response from a single HTML read reduces the chance of encountering a HTTP 503 error when parsing multiple sections from a single URL. If a response argument is missing, a response is read. Followed by the `sleep_time` in seconds.}

\item{schema}{Schema that the returned data is written to, Default: 'chemidplus'}

\item{sleep_time}{If the response argument is missing, the number seconds to pause after reading the URL, Default: 3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[xml2]{read_xml}}
 \code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
 \code{\link[stringr]{str_remove}}
 \code{\link[centipede]{strsplit}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{distinct}}
 \code{\link[purrr]{map2}},\code{\link[purrr]{set_names}},\code{\link[purrr]{map}}
 \code{\link[tibble]{as_tibble}}
}
