% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R
\name{get_ncit_synonym}
\alias{get_ncit_synonym}
\title{Scrape the NCI Thesaurus}
\usage{
get_ncit_synonym(
  conn,
  sleep_time = 5,
  expiration_days = 30,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{sleep_time}{Time in seconds for the system to sleep before each scrape with \code{\link[xml2]{read_html}}.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\description{
All NCIt Codes that have not been scraped or were scraped in the expiration period are scraped in the NCIt Thesaurus at the \url{"https://ncithesaurus.nci.nih.gov/ncitbrowser/pages/concept_details.jsf?dictionary=NCI_Thesaurus&code=\%s&ns=ncit&type=synonym&key=null&b=1&n=0&vse=null#} path.
}
\section{Drug Detail Links}{

The links to Drug Pages are scraped from the Data Dictionary URL over the maximum page number and are saved to a Drug Link Table in the \code{cancergov} schema. The URLs in the Drug Link Table are then scraped for any HTML Tables of synonyms and the results are written to a Drug Link Synonym Table. The links to active clinical trials and NCIt mappings are also derived and stored in their respective tables.
}

\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[SqlRender]{render}}
\code{\link[secretary]{typewrite_progress}},\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{character(0)}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[purrr]{keep}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select_all}}
\code{\link[rubix]{format_colnames}}
}
