% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-internals.R
\name{get_pm_earliest}
\alias{get_pm_earliest}
\title{Get Earliest PubMed Publications}
\usage{
get_pm_earliest(conn, conn_fun, search_term, verbose = TRUE, render_sql = TRUE)
}
\description{
Get the earliest PubMed publications for a given term. The results are stored in the PM_EARLIEST Table. See \code{\link{get_pm}} for more details.
}
\seealso{
Other pubmed functions: 
\code{\link{get_pm_latest}()},
\code{\link{get_pm}()}
}
\concept{pubmed functions}
