% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_functions.R
\name{get_pm_latest}
\alias{get_pm_latest}
\title{Get Latest PubMed Publications}
\usage{
get_pm_latest(conn, conn_fun, search_term, verbose = TRUE, render_sql = TRUE)
}
\description{
Get the latest PubMed publications for a given term. The results are stored in the PM_LATEST Table. See \code{\link{get_pm}} for more details.
}
\seealso{
Other pubmed functions: 
\code{\link{get_pm_earliest}()},
\code{\link{get_pm}()}
}
\concept{pubmed functions}
