% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_functions.R
\name{get_response}
\alias{get_response}
\title{Get a Response from a RN URL}
\usage{
get_response(rn_url, sleep_time = 3)
}
\arguments{
\item{rn_url}{PARAM_DESCRIPTION}

\item{sleep_time}{PARAM_DESCRIPTION, Default: 3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xml2]{read_xml}}
}
