% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rn_url_validity.R
\name{get_rn_url_validity}
\alias{get_rn_url_validity}
\title{Check if the RN URL derived from the Registry Number is valid}
\usage{
get_rn_url_validity(
  conn,
  rn_url,
  response,
  schema = "chemidplus",
  sleep_time = 3
)
}
\description{
Check if the RN URL derived from the Registry Number is valid
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate}}
 \code{\link[stringr]{str_remove}}
 \code{\link[xml2]{read_xml}}
 \code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
}
