% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugDictionaryConcept.R
\name{getDrugDictionaryConcept}
\alias{getDrugDictionaryConcept}
\title{Get Drug Dictionary Data}
\usage{
getDrugDictionaryConcept(conn, max_page)
}
\description{
Scrape, Load, and Format in the OMOP Vocabulary Concept Table format.
}
