% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancergov_functions.R
\name{has_new_drug_count}
\alias{has_new_drug_count}
\title{FUNCTION_TITLE}
\usage{
has_new_drug_count(conn)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
This function must be run before a log entry is made using the log_drug_count function.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}}
\code{\link[skyscraper]{nci_count}}
}
