% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is404.R
\name{is404}
\alias{is404}
\title{Is the RN URL returning an HTTP error 404?}
\usage{
is404(rn_url)
}
\arguments{
\item{rn_url}{Registry Number URL}
}
\value{
logical vector of length 1
}
\description{
This function gets a response from an RN URL and checks for a 404 message in the error output.
}
\seealso{
\code{\link[xml2]{read_xml}}

Other chemidplus parsing: 
\code{\link{isSingleHit}()}
}
\concept{chemidplus parsing}
