% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMultipleHits.R
\name{isMultipleHits}
\alias{isMultipleHits}
\title{Get the RN Number an URL for a Phrase}
\usage{
isMultipleHits(response)
}
\description{
Get the RN Number an URL for a Phrase
}
\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
 \code{\link[tibble]{as_tibble}}
 \code{\link[rubix]{filter_at_grepl}}
 \code{\link[tidyr]{extract}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{mutate_all}}
 \code{\link[stringr]{str_remove}}
}
