% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_functions.R
\name{isMultipleHits3}
\alias{isMultipleHits3}
\title{FUNCTION_TITLE}
\usage{
isMultipleHits3(response)
}
\arguments{
\item{response}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[stringr]{str_remove}},\code{\link[stringr]{modifiers}}
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{bind}}
}
