% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNoRecord.R
\name{isNoRecord}
\alias{isNoRecord}
\title{Does the URL return no records found?}
\usage{
isNoRecord(response)
}
\arguments{
\item{response}{xml_document xml_node class object}
}
\description{
Does the URL return no records found?
}
\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
}
