% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_functions.R
\name{isSingleHit2}
\alias{isSingleHit2}
\title{FUNCTION_TITLE}
\usage{
isSingleHit2(response)
}
\arguments{
\item{response}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[centipede]{strsplit}},\code{\link[centipede]{no_na}}
\code{\link[tibble]{as_tibble}},\code{\link[tibble]{tribble}}
\code{\link[tidyr]{extract}},\code{\link[tidyr]{pivot_wider}}
\code{\link[dplyr]{mutate}}
\code{\link[stringr]{str_remove}}
}
