% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_to_omop_functions.R
\name{join_chemidplus_omop_synonyms}
\alias{join_chemidplus_omop_synonyms}
\title{Join ChemiDPlus Results to the OMOP Vocabulary by Concept}
\usage{
join_chemidplus_omop_synonyms(
  conn,
  chemidplusSchema,
  vocabularySchema = "public",
  vocabularyIds = c("HemOnc", "RxNorm", "RxNorm Extension", "ATC")
)
}
\arguments{
\item{conn}{Postgres connection}

\item{chemidplusSchema}{Schema that contains the Registry Number Log and Names and Synonyms` Tables}

\item{vocabularySchema}{Schema that contains the OMOP Concept and Concept Synonym Tables, Default: 'public'}

\item{vocabularyIds}{Vocabulary Ids in OMOP Concept Table to filter results for. Default: c("HemOnc", "RxNorm", "RxNorm Extension", "ATC")}
}
\value{
Dataframe with all the Raw Concepts from the Registry Number Log with a Registry Number,  the Timestamp of when the Names and Synonyms Table was appended, Registry Number URL, the Synonym Type and Synonym scraped from the URL, and all the fields found in the OMOP Concept Table that the Synonym mapped to via any of the OMOP Concept Synonyms (not returned).
}
\description{
All the Names and Synonyms for a given Raw Concept are joined in lowercase to the lowercase Concept Synonyms found in the OMOP Concept Synonym Table that are valid and belong to the Drug domain.
}
\seealso{
\code{\link[pg13]{query}}
\code{\link[SqlRender]{render}}
}
