% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_mth.R, R/supplementary_functions.R
\name{load_all_mth}
\alias{load_all_mth}
\alias{load_mrconso}
\title{FUNCTION_TITLE}
\usage{
load_all_mth(umls_conn, conn)

load_mrconso(umls_conn, conn)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
https://evs.nci.nih.gov/evs-download/thesaurus-downloads

https://evs.nci.nih.gov/evs-download/thesaurus-downloads
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{send}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}

\code{\link[pg13]{send}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}

Other nci evs schema: 
\code{\link{load_ncit}()}

Other nci evs schema: 
\code{\link{load_ncit}()}
}
\concept{nci evs schema}
