% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCachedDrugPage.R
\name{loadCachedDrugPage}
\alias{loadCachedDrugPage}
\title{Load Cached Drug Page}
\usage{
loadCachedDrugPage(df, output.var = "loadCachedDrugPage_results")
}
\arguments{
\item{df}{Data Frame of any dimensions with a 'Drug' and 'Drug_Def_Link' fields that correspond to the drug and its hyperlink}

\item{output.var}{Name of output list from all the loaded cached results, Default: 'loadCachedDrugPage_results'}
}
\value{
Named list of all the cache results by drug link.
}
\description{
Regardless of whether or not the scraping has been completed, load all cached Drug Page scrapes.
}
\details{
This function bookmarks the process if it is stopped before completion. If a complete reset is desired, either delete the output.var object in the Global Environment or state "n" at prompt "output.var already exists. Continue? ". All other values indicate continuing from the length of the output.var.
}
\seealso{
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{bind}}
 \code{\link[secretary]{typewrite}},\code{\link[secretary]{typewrite_bold}}
}
