% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadScrapedDrugDict.R
\name{loadScrapedDrugDict}
\alias{loadScrapedDrugDict}
\title{Load Scraped Cancer.gov Drug Dictionary Data}
\usage{
loadScrapedDrugDict(max_page = 39)
}
\description{
Load Scraped Cancer.gov Drug Dictionary Data
}
