% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadScrapedDrugs.R
\name{loadScrapedDrugs}
\alias{loadScrapedDrugs}
\title{Scrape Drug Page}
\usage{
loadScrapedDrugs(.input, starting_row, ending_row)
}
\arguments{
\item{max_page}{maximum page for the base url https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=}
}
\description{
Scrape Drug Page
}
