% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R
\name{log_drug_count}
\alias{log_drug_count}
\title{Log the Drug Count in the Drug Dictionary}
\usage{
log_drug_count(conn, verbose = TRUE, render_sql = TRUE)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\description{
Log the drug count scraped by \code{\link{drug_count}} with a timestamp to the Drug Dictionary Log Table and also receive a time difference and count comparison between the most recent count and the current count in the R console.
}
\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[secretary]{typewrite}}
\code{\link[tibble]{tibble}}
}
