% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdp-internals.R
\name{log_errors}
\alias{log_errors}
\title{FUNCTION_TITLE}
\usage{
log_errors(conn, schema, errors, search_type)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}

\item{errors}{PARAM_DESCRIPTION}

\item{search_type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{mutate}}
\code{\link[stringr]{str_remove}}
\code{\link[pg13]{dropTable}},\code{\link[pg13]{writeTable}},\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
}
