% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_registry_number.R
\name{log_registry_number}
\alias{log_registry_number}
\title{Get the Registry Number for a Given String}
\usage{
log_registry_number(
  conn,
  raw_concept,
  type = "contains",
  sleep_time = 3,
  schema = "chemidplus"
)
}
\description{
Get the Registry Number for a Given String
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate}}
 \code{\link[stringr]{str_remove}}
 \code{\link[xml2]{read_xml}}
 \code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
}
