% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_umls_api.R
\name{log_umls_search}
\alias{log_umls_search}
\title{FUNCTION_TITLE}
\usage{
log_umls_search(conn, string, schema = "omop_drug_to_umls_api")
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{string}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_drug_to_umls_api'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}},\code{\link[pg13]{send}},\code{\link[pg13]{appendTable}}
\code{\link[dplyr]{distinct}}
}
