% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-methods.R
\name{lookup_ncit_code}
\alias{lookup_ncit_code}
\title{Lookup an NCIt Code}
\usage{
lookup_ncit_code(
  ncit_code,
  conn,
  sleep_time = 5,
  verbose = TRUE,
  render_sql = TRUE,
  encoding = "",
  options = c("RECOVER", "NOERROR", "NOBLANKS")
)
}
\arguments{
\item{ncit_code}{NCI Thesaurus Code to lookup.}

\item{conn}{Postgres connection object.}

\item{sleep_time}{Time in seconds for the system to sleep before each scrape with \code{\link[xml2]{read_html}}.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}

\item{encoding}{Specify a default encoding for the document. Unless
otherwise specified XML documents are assumed to be in UTF-8 or
UTF-16. If the document is not UTF-8/16, and lacks an explicit
encoding directive, this allows you to supply a default.}

\item{options}{Set parsing options for the libxml2 parser. Zero or more of
\describe{
  \item{RECOVER}{recover on errors}
  \item{NOENT}{substitute entities}
  \item{DTDLOAD}{load the external subset}
  \item{DTDATTR}{default DTD attributes}
  \item{DTDVALID}{validate with the DTD}
  \item{NOERROR}{suppress error reports}
  \item{NOWARNING}{suppress warning reports}
  \item{PEDANTIC}{pedantic error reporting}
  \item{NOBLANKS}{remove blank nodes}
  \item{SAX1}{use the SAX1 interface internally}
  \item{XINCLUDE}{Implement XInclude substitition}
  \item{NONET}{Forbid network access}
  \item{NODICT}{Do not reuse the context dictionary}
  \item{NSCLEAN}{remove redundant namespaces declarations}
  \item{NOCDATA}{merge CDATA as text nodes}
  \item{NOXINCNODE}{do not generate XINCLUDE START/END nodes}
  \item{COMPACT}{compact small text nodes; no modification of the tree allowed afterwards (will possibly crash if you try to modify the tree)}
  \item{OLD10}{parse using XML-1.0 before update 5}
  \item{NOBASEFIX}{do not fixup XINCLUDE xml:base uris}
  \item{HUGE}{relax any hardcoded limit from the parser}
  \item{OLDSAX}{parse using SAX2 interface before 2.7.0}
  \item{IGNORE_ENC}{ignore internal document encoding hint}
  \item{BIG_LINES}{Store big lines numbers in text PSVI field}
}}
}
\description{
Lookup an NCIt Code in the NCIt Synonym Table. If the resultset has 0 rows, a scrape of the url path to the NCIt entry is done and any parsed response is saved to the Table.
}
\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[SqlRender]{render}}
\code{\link[secretary]{typewrite}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[purrr]{keep}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select_all}}
\code{\link[rubix]{format_colnames}}
}
