% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancergov_functions.R
\name{lookup_ncit_synonym}
\alias{lookup_ncit_synonym}
\title{FUNCTION_TITLE}
\usage{
lookup_ncit_synonym(ncit_code, conn, sleep_time = 5)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{sleep_time}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{readTable}},\code{\link[pg13]{appendTable}}
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[purrr]{keep}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select_all}}
\code{\link[rubix]{format_colnames}}
}
