% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maintenance_functions.R
\name{map_schema}
\alias{map_schema}
\title{Map a skyscraper Schema to a Data Package Repository}
\usage{
map_schema(repo_username = "meerapatelmd")
}
\arguments{
\item{repo_username}{Github username to which the Data Package belongs, Default: 'meerapatelmd'}
}
\description{
Each Postgres schema governed by this package has a corresponding Data Package, where updates to the Postgres schema are exported as csvs, packaged, and pushed to a GitHub repo.
}
\seealso{
\code{\link[tibble]{tribble}}
\code{\link[dplyr]{mutate}}

Other local maintenance: 
\code{\link{export_schema_to_data_repo}()},
\code{\link{import_schemas}()},
\code{\link{schema_status}()}
}
\concept{local maintenance}
