% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nci_count.R
\name{nci_count}
\alias{nci_count}
\title{Get the NCI Drug Dictionary Count}
\usage{
nci_count()
}
\value{
An integer in the "X results found for: ALL" at "https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=1"
}
\description{
Get the total number of drugs in the NCI Drug Dictionary
}
\seealso{
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
}
