% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nci_dd-internals.R
\name{nci_internal}
\alias{nci_internal}
\title{NCI Drug Dictionary Internal Functions}
\description{
This is an internal function to the \verb{_run*} function, which is part of the family of functions that scrape, parse, and store the NCI Drug Dictionary found at CancerGov.org and any correlates to the NCI Thesaurus in a Postgres Database. Use \code{\link{cg_run}} to run the full sequence. See details for more info.
}
