% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nciDrugCount.R
\name{nciDrugCount}
\alias{nciDrugCount}
\title{Scrape Cancer.gov Drug Dictionary}
\usage{
nciDrugCount()
}
\arguments{
\item{max_page}{Maximum page to scrape until at https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=}
}
\description{
Scrape Cancer.gov Drug Dictionary
}
