% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nciDrugDetailLinks.R
\name{nciDrugDetailLinks}
\alias{nciDrugDetailLinks}
\title{Get the Links to Drug Pages}
\usage{
nciDrugDetailLinks(max_page = 39)
}
\arguments{
\item{max_page}{maximum page for the base url https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=}
}
\description{
Get the Links to Drug Pages
}
