% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{onLoad}
\alias{onLoad}
\title{Setup and Update skyscraper Schemas}
\usage{
onLoad(schemas, conn)
}
\arguments{
\item{schemas}{(optional) character vector of 1 or more schemas to filter for setup and updating}
}
\description{
Setup and Update skyscraper Schemas
}
\seealso{
\code{\link[rubix]{filter_for}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[utils]{capture.output}}
 \code{\link[devtools]{remote-reexports}}
 \code{\link[dplyr]{filter}}
}
