% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R
\name{process_drug_link_ncit}
\alias{process_drug_link_ncit}
\title{Process the NCIt CUI from the Drug Link URL Table}
\usage{
process_drug_link_ncit(
  conn,
  verbose = TRUE,
  render_sql = TRUE,
  expiration_days = 30
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\description{
Run the full sequence that scrapes, parses, and stores the NCI Drug Dictionary found at CancerGov.org and any correlates to the NCI Thesaurus in a Postgres Database.
}
\details{
Unlike the other \verb{process_*} functions, this one uses regex to find the NCI Thesaurus Code, if presence, from the scraped URLs instead of performing any scraping.
}
\section{Drug Detail Links}{

The links to Drug Pages are scraped from the Data Dictionary URL over the maximum page number and are saved to a Drug Link Table in the \code{cancergov} schema. The URLs in the Drug Link Table are then scraped for any HTML Tables of synonyms and the results are written to a Drug Link Synonym Table. The links to active clinical trials and NCIt mappings are also derived and stored in their respective tables.
}

\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[SqlRender]{render}}
\code{\link[rubix]{filter_at_grepl}}
\code{\link[tidyr]{extract}}
\code{\link[dplyr]{mutate}}
}
