% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R
\name{process_drug_link_synonym}
\alias{process_drug_link_synonym}
\title{Process the Links found in the Drug Link Table for Synonyms}
\usage{
process_drug_link_synonym(
  conn,
  sleep_time = 3,
  expiration_days = 30,
  verbose = TRUE,
  render_sql = TRUE,
  encoding = "",
  options = c("RECOVER", "NOERROR", "NOBLANKS")
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{sleep_time}{Time in seconds for the system to sleep before each scrape with \code{\link[xml2]{read_html}}.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\description{
Run the full sequence that scrapes, parses, and stores the NCI Drug Dictionary found at CancerGov.org and any correlates to the NCI Thesaurus in a Postgres Database.
}
\section{Drug Detail Links}{

The links to Drug Pages are scraped from the Data Dictionary URL over the maximum page number and are saved to a Drug Link Table in the \code{cancergov} schema. The URLs in the Drug Link Table are then scraped for any HTML Tables of synonyms and the results are written to a Drug Link Synonym Table. The links to active clinical trials and NCIt mappings are also derived and stored in their respective tables.
}

\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}}
\code{\link[SqlRender]{render}}
\code{\link[secretary]{typewrite_progress}},\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{character(0)}}
\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_replace}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter_all}}
\code{\link[tidyr]{separate_rows}}
\code{\link[tibble]{tibble}}
}
