% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancergov_functions.R
\name{process_drug_link_url}
\alias{process_drug_link_url}
\title{Process the Links found in the Drug Link Table for NCIt and other URLs}
\usage{
process_drug_link_url(conn, sleep_time = 5)
}
\arguments{
\item{conn}{Postgres connection object}

\item{sleep_time}{Seconds in between xml2::read_html calls on a URL, Default: 5}
}
\description{
These functions scrape and parse the NCI Drug Dictionary found at CancerGov.org.
}
\section{Web Source Types}{

The NCI Drug Dictionary has 2 data sources that run in parallel. The first source is the Drug Dictionary itself at \href{https://www.cancer.gov/publications/dictionaries/cancer-drug}{https://www.cancer.gov/publications/dictionaries/cancer-drug}. The other source are the individual drug pages, called Drug Detail Links in skyscraper, that contain tables of synonyms, including investigational names.
}

\section{Drug Detail Links}{

The links to Drug Pages are scraped from the Data Dictionary URL over the maximum page number and are saved to a Drug Link Table in the \code{cancergov} schema. The URLs in the Drug Link Table are then scraped for any HTML Tables of synonyms and the results are written to a Drug Link Synonym Table. The links to active clinical trials and NCIt mappings are also derived and stored in their respective tables.
}

\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}},\code{\link[pg13]{readTable}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter_all}}
\code{\link[police]{try_catch_error_as_null}}
\code{\link[xml2]{read_xml}},\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_replace}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[tidyr]{separate_rows}}

Other cancergov: 
\code{\link{get_dictionary_and_links}()},
\code{\link{get_drug_detail_links}()},
\code{\link{get_drug_dictionary}()},
\code{\link{get_drug_link_synonym}()},
\code{\link{get_drug_link_url}()},
\code{\link{log_drug_count}()},
\code{\link{process_drug_link_ncit}()},
\code{\link{process_drug_link_synonym}()}
}
\concept{cancergov}
