% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maintenance_functions.R
\name{schema_status}
\alias{schema_status}
\title{Get the Status of a Data Package Installation}
\usage{
schema_status()
}
\description{
Local maintenance functions are meant to be used to export and instantiate/update a local skyscraper schema with the Data Package Repo. This is important for keeping data updated on one machine when there have been additions made to the Remote Data Package Repository.
}
\section{Check Schema Status}{

All the Data Packages in the Schema Map returned by \code{\link{map_schema}} are installed if there are changes to the SHA hash. An "updated" field is added to the Schema Map that contain the values of TRUE if a new installation occurred and FALSE otherwise. If an update to the schema is desired after running this function, \code{\link{update_schemas}}, should be run with \code{force_update} set to \code{TRUE}.
}

\seealso{
\code{\link[rubix]{filter_for}}
\code{\link[purrr]{map}}
\code{\link[utils]{capture.output}}
\code{\link[devtools]{remote-reexports}}

Other local maintenance: 
\code{\link{export_schema_to_data_repo}()},
\code{\link{import_schemas}()},
\code{\link{map_schema}()}
}
\concept{local maintenance}
