% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaMap.R
\name{schemaMap}
\alias{schemaMap}
\title{Map a skyscraper Postgres Schema to a Data Package}
\usage{
schemaMap()
}
\description{
Map a skyscraper Postgres Schema to a Data Package
}
\seealso{
\code{\link[tibble]{tribble}}
}
