% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_functions.R
\name{scrape_latest_pubmed}
\alias{scrape_latest_pubmed}
\title{FUNCTION_TITLE}
\usage{
scrape_latest_pubmed(conn, search_term, schema = "patelm9")
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{search_term}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION, Default: 'patelm9'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_remove}}
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}},\code{\link[pg13]{send}}
\code{\link[SqlRender]{render}}
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[tibble]{tibble}}
\code{\link[tidyr]{extract}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
\code{\link[lubridate]{parse_date_time}}
}
