% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeCancerGovDict.R
\name{scrapeCancerGovDict}
\alias{scrapeCancerGovDict}
\title{Scrape Cancer.gov Drug Dictionary}
\usage{
scrapeCancerGovDict(max_page = 39)
}
\arguments{
\item{max_page}{Maximum page to scrape until at https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=}
}
\description{
Scrape Cancer.gov Drug Dictionary
}
