% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeDrugPages.R
\name{scrapeDrugPages}
\alias{scrapeDrugPages}
\title{Scrape Drug Page}
\usage{
scrapeDrugPages(max_page)
}
\arguments{
\item{max_page}{maximum page for the base url https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=}
}
\description{
Scrape Drug Page
}
