% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchChemiDPlus.R
\name{searchChemiDPlus}
\alias{searchChemiDPlus}
\title{Search ChemiDPlus and Store Results in Tables}
\usage{
searchChemiDPlus(
  conn,
  search_term,
  type = "contains",
  sleep_time = 3,
  schema = "chemidplus_search",
  export_repo = TRUE,
  target_dir = "~/GitHub/Public-Packages/chemidplusSearchData/"
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{search_term}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: 'contains'}

\item{sleep_time}{PARAM_DESCRIPTION, Default: 3}

\item{schema}{PARAM_DESCRIPTION, Default: 'chemidplus_search'}

\item{export_repo}{PARAM_DESCRIPTION, Default: TRUE}

\item{target_dir}{PARAM_DESCRIPTION, Default: '~/GitHub/Public-Packages/chemidplusSearchData/'}
}
\description{
Search ChemiDPlus and Store Results in Tables
}
\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{filter}}
 \code{\link[centipede]{no_na}}
 \code{\link[tibble]{tribble}},\code{\link[tibble]{c("tibble", "tibble")}}
 \code{\link[police]{try_catch_error_as_null}}
}
