% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupChemiDPlus.R
\name{setupChemiDPlus}
\alias{setupChemiDPlus}
\title{Setup the ChemiDPlus Schema}
\usage{
setupChemiDPlus(conn)
}
\arguments{
\item{conn}{Postgres connection}
}
\description{
If a `chemidplus` schema does not exist, a `chemidplus` Schema at a given Postgres connection with the data found in the chemidplusData Package. If the `chemidplus` Schema needs to be updated with new data added to the chemidplusData Package, use the updateChemiDPlus function instead.
}
\details{
The chemidplusData Package is force installed first. if a `chemidplus` schema is not present in the given connection,`chemidplus` schema is created, and the data exported in the chemidplusData R Package is written as tables.
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{writeTable}}
 \code{\link[devtools]{remote-reexports}}
 \code{\link[rubix]{map_names_set}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[dplyr]{mutate_all}}
 \code{\link[lubridate]{ymd_hms}}
}
