% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRNtoOMOP.R
\name{setupRNtoOMOP}
\alias{setupRNtoOMOP}
\title{Transform the RN Tables into OMOP Vocabulary Tables}
\usage{
setupRNtoOMOP(conn)
}
\arguments{
\item{conn}{Postgres connection}
}
\description{
This function writes the first Concept, Concept Synonym, Concept Ancestor, and Concept Relationship Tables from the Classification, Synonyms, and Phrase Log Tables in the chemidplus schema. If the Concept, Concept Synonym, Concept Ancestor, and/or Concept Relationship Tables already exist in the chemidplus Schema, an error will be thrown.
}
\details{
concept_class_id is the search type. For broader search types such as "contains", there will be the same synonyms matched to different concept ids such as 35101002 & 35100499
}
\seealso{
\code{\link[pg13]{readTable}},\code{\link[pg13]{query}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
 \code{\link[rubix]{mutate_all_as_char}},\code{\link[rubix]{mutate_all_na_str_to_na}},\code{\link[rubix]{make_identifier}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{group_by}}
 \code{\link[tidyr]{extract}}
 \code{\link[chariot]{queryAthena}},\code{\link[chariot]{filterValid}}
 \code{\link[stringr]{str_remove}}
}
