% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupSchema.R
\name{setupSchema}
\alias{setupSchema}
\title{Setup a Schema}
\usage{
setupSchema(conn, schema = "chemidplus", dataPackage = "chemidplusData")
}
\arguments{
\item{conn}{Postgres connection}
}
\description{
If the schema does not exist, the schema at a given Postgres connection with the data found in the chemidplusData Package. If the schema needs to be updated with new data added to the corresponding data Package, use the updateChemiDPlus function instead.
}
\details{
The chemidplusData Package is force installed first. if a `chemidplus` schema is not present in the given connection,`chemidplus` schema is created, and the data exported in the chemidplusData R Package is written as tables.
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{writeTable}}
 \code{\link[devtools]{remote-reexports}}
 \code{\link[rubix]{map_names_set}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[dplyr]{mutate_all}}
 \code{\link[lubridate]{ymd_hms}}
}
