% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyscraper-package.R
\docType{package}
\name{skyscraper-package}
\alias{skyscraper}
\alias{skyscraper-package}
\title{skyscraper: Scrape Clinical Drug Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Scrape, store, and maintain clinical drug data
    across various disparate web and data sources, including the
    ChemiDPlus, NCI Drug Dictionary, and the ClinicalTrials.gov AACT
    Database.
}
\section{Web Sources}{


ChemiDPlus: \href{https://chem.nlm.nih.gov/chemidplus/}{https://chem.nlm.nih.gov/chemidplus/}

NCI Drug Dictionary: \href{https://www.cancer.gov/publications/dictionaries/}{https://www.cancer.gov/publications/dictionaries/}

UpToDate

PubMed Search Results
}

\section{Data Sources}{

ClinicalTrials.gov
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meerapatelmd/skyscraper}
  \item \url{https://meerapatelmd.github.io/skyscraper}
  \item Report bugs at \url{https://github.com/meerapatelmd/skyscraper/issues}
}

}
\author{
\strong{Maintainer}: Meera Patel, MD \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
