% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyscraper.R
\docType{data}
\name{skyscraper}
\alias{skyscraper}
\title{Skyscraper data}
\format{A tbl_df with 100 rows and 9 variables :
\itemize{
\item rank (int)
\item building_name (chr)
\item city (fctr)
\item height_m (dbl)
\item height_ft (dbl)
\item floors (int)
\item year_completed (int)
\item material (fctr)
\item use (chr)
}}
\source{
\url{http://www.skyscrapercenter.com/buildings}
}
\usage{
skyscraper
}
\description{
100 Tallest Completed Buildings in the World (as of 2015-11-25)
}
\examples{
str(skyscraper)
head(skyscraper)
plot(height_m ~ year_completed, skyscraper, subset = year_completed > 2005)
}
\keyword{datasets}

