% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-utils.R
\name{start_pm}
\alias{start_pm}
\title{Create PubMed Tables in Patelm9 Schema}
\usage{
start_pm(conn, verbose = TRUE, render_sql = TRUE)
}
\description{
Create PubMed Tables in the Patelm9 schema if they do not already exist. Tables include: PM_LOG, PM_EARLIEST, PM_LATEST, and PM_RESULTS*. PM_RESULTS is not created by this script. It is written by \code{\link{pm_run}}.
}
