% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R
\name{update_cancergov_drugs}
\alias{update_cancergov_drugs}
\title{Update the Cancergov Drugs Table}
\usage{
update_cancergov_drugs(conn, verbose = TRUE, render_sql = TRUE)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\description{
The Synonyms found in the NCI Drug Dictionary and NCI Thesaurus are aggregated into a single Cancergov Drug Table.
}
\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{send}}
\code{\link[tidyr]{pivot_longer}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{coalesce}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{reexports}}
\code{\link[forcats]{fct_recode}}
}
