% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateChemiDPlus.R
\name{updateChemiDPlus}
\alias{updateChemiDPlus}
\title{Update the ChemiDPlus Schema with New Data}
\usage{
updateChemiDPlus(conn)
}
\description{
If the current installation of the chemidplusData Package is not up-to-date, it is installed and the current `chemidplus` schema is dropped and a new schema is written with the fresh data.
}
\seealso{
\code{\link[utils]{capture.output}}
 \code{\link[devtools]{remote-reexports}}
 \code{\link[pg13]{dropSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{writeTable}}
 \code{\link[rubix]{map_names_set}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[dplyr]{mutate_all}}
 \code{\link[lubridate]{ymd_hms}}
}
