% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateClinicalTrialsGov.R
\name{updateClinicalTrialsGov}
\alias{updateClinicalTrialsGov}
\title{Update Clinical Trial Gov Database}
\usage{
updateClinicalTrialsGov()
}
\description{
Update Clinical Trial Gov Database
}
\seealso{
\code{\link[xml2]{read_xml}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
}
