% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-utils.R
\name{write_cg_staging_tbl}
\alias{write_cg_staging_tbl}
\title{Write Staging Tables to Cancergov Schema}
\usage{
write_cg_staging_tbl(conn, tableName, data, verbose = TRUE, render_sql = TRUE)
}
\description{
Write Staging Tables to Cancergov Schema
}
