% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheScrape}
\alias{cacheScrape}
\title{Cache the Scraped object}
\usage{
cacheScrape(object, url)
}
\arguments{
\item{source}{Source vocabulary such as "CancerGov" or "UpToDate"}
}
\description{
Cache the Scraped object
}
