% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdp.R
\name{cdp_run}
\alias{cdp_run}
\title{Search ChemiDPlus and Store Results}
\usage{
cdp_run(
  conn,
  steps = c("log_registry_number", "get_rn_url_validity", "get_classification",
    "get_names_and_synonyms", "get_registry_numbers", "get_links_to_resources"),
  search_term,
  search_type = "contains",
  sleep_time = 5,
  verbose = TRUE,
  render_sql = TRUE
)
}
\description{
Search ChemiDPlus and Store Results
}
\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{filter}}
\code{\link[tibble]{tribble}},\code{\link[tibble]{c("tibble", "tibble")}}
\code{\link[police]{try_catch_error_as_null}}
}
