% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddlCancerGovSchema.R
\name{ddlCancerGovSchema}
\alias{ddlCancerGovSchema}
\title{DDL Standardized Vocabulary Tables}
\usage{
ddlCancerGovSchema(conn)
}
\description{
The ddl will only execute on the condition that there are 0 tables current in the `hemonc_extension` schema.
}
