% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg-internals.R, R/nci_dd-internals.R
\name{drug_count}
\alias{drug_count}
\alias{get_nci_dd}
\title{Get the Drug Count in the Drug Dictionary}
\usage{
drug_count(size = 10000, crawl_delay = 5)

get_nci_dd(crawl_delay = 5, size = 10000, letters, verbose = TRUE)

drug_count(size = 10000, crawl_delay = 5)
}
\arguments{
\item{size}{The number of records to retrieve.}

\item{crawl_delay}{Delay in seconds.}
}
\value{
Drug count as integer

A list with results and metadata.

Drug count as integer
}
\description{
This is an internal function to the \verb{_run*} function, which is part of the family of functions that scrape, parse, and store the NCI Drug Dictionary found at CancerGov.org and any correlates to the NCI Thesaurus in a Postgres Database. Use \code{\link{cg_run}} to run the full sequence. See details for more info.
}
\details{
Retrieve the total number of drugs in the NCI Drug Dictionary from the Drug Dictionary API (\url{https://webapis.cancer.gov/drugdictionary/v1/index.html#/Drugs/Drugs_GetByName})

Retrieve the total number of drugs in the NCI Drug Dictionary from the Drug Dictionary API (\url{https://webapis.cancer.gov/drugdictionary/v1/index.html#/Drugs/Drugs_GetByName})

Retrieve the total number of drugs in the NCI Drug Dictionary from the Drug Dictionary API (\url{https://webapis.cancer.gov/drugdictionary/v1/index.html#/Drugs/Drugs_GetByName})
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
\code{\link[jsonlite]{toJSON, fromJSON}}

\code{\link[httr]{GET}},\code{\link[httr]{content}}
\code{\link[jsonlite]{toJSON, fromJSON}}

\code{\link[httr]{GET}},\code{\link[httr]{content}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
