% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancergov_functions.R
\name{get_dictionary_and_links}
\alias{get_dictionary_and_links}
\title{Scrape the Drug Definitions and Links from the NCI Drug Dictionary}
\usage{
get_dictionary_and_links(
  conn,
  max_page = 50,
  sleep_time = 3,
  progress_bar = TRUE
)
}
\arguments{
\item{conn}{Postgres connection object}

\item{max_page}{maximum page number to iterate the scrape over in the "https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=" path, Default: 50

.}
}
\value{
Drug Dictionary and Drug Link Table in the \code{cancergov} schema if a Drug Dictionary Table didn't already exist. Otherwise, both Tables are appended with any new observations scraped.
}
\description{
These functions scrape and parse the NCI Drug Dictionary found at CancerGov.org.
}
\details{
This function combines the operations of \code{\link{get_drug_detail_links}} and \code{\link{get_drug_dictionary}} to parse from a single response.
}
\section{Web Source Types}{

The NCI Drug Dictionary has 2 data sources that run in parallel. The first source is the Drug Dictionary itself at \href{https://www.cancer.gov/publications/dictionaries/cancer-drug}{https://www.cancer.gov/publications/dictionaries/cancer-drug}. The other source are the individual drug pages, called Drug Detail Links in skyscraper, that contain tables of synonyms, including investigational names.
}

\section{Drug Dictionary}{

The listed drug names and their definitions are scraped from the Drug Dictionary HTML and updated to a Drug Dictionary Table in a \code{cancergov} schema.
}

\seealso{
\code{\link[tibble]{tibble}}
\code{\link[secretary]{typewrite}}
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[purrr]{map}},\code{\link[purrr]{keep}},\code{\link[purrr]{transpose}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}}
\code{\link[stringr]{str_replace}}

Other cancergov: 
\code{\link{get_drug_detail_links}()},
\code{\link{get_drug_dictionary}()},
\code{\link{get_drug_link_synonym}()},
\code{\link{get_drug_link_url}()},
\code{\link{log_drug_count}()},
\code{\link{process_drug_link_ncit}()},
\code{\link{process_drug_link_synonym}()}
}
\concept{cancergov}
