% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drug_detail_links.R
\name{get_drug_detail_links}
\alias{get_drug_detail_links}
\title{Get the Links to Drug Pages}
\usage{
get_drug_detail_links(conn, max_page = 50)
}
\arguments{
\item{max_page}{maximum page for the base url https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=}
}
\description{
Get the Links to Drug Pages
}
