% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_functions.R
\name{isNoRecord}
\alias{isNoRecord}
\title{Does the RN URL indicate that no records were found?}
\usage{
isNoRecord(response)
}
\arguments{
\item{response}{"xml_document" "xml_node" class object returned by calling xml2::read_html on the rn_url}
}
\description{
There are 2 HTML responses involved in ChemiDPlus scraping. The first HTML response is associated with the concept search while the second is for the 1 to 5 Registry Numbers and their URLs if records were found. The ChemiDPlus parsing functions delineate the response into 3 groups: searches the resulted in no records, a single record, or multiple records (currently limited to parsing a maximum of 5 records).
}
\section{No Records}{

If the ChemiDPlus search url response contained the phrase "The following query produced no records:", it is marked as not having any records.
}

\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}

Other chemidplus parsing: 
\code{\link{is404}()},
\code{\link{isMultipleHits}()},
\code{\link{isSingleHit}()}
}
\concept{chemidplus parsing}
