% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdp-utils.R
\name{isSingleHit}
\alias{isSingleHit}
\title{Parse the RN from a single Substance Page}
\usage{
isSingleHit(response)
}
\arguments{
\item{response}{"xml_document" "xml_node" class object returned by calling xml2::read_html on the rn_url}
}
\description{
There are 2 HTML responses involved in ChemiDPlus scraping. The first HTML response is associated with the concept search while the second is for the 1 to 5 Registry Numbers and their URLs if records were found. The ChemiDPlus parsing functions delineate the response into 3 groups: searches the resulted in no records, a single record, or multiple records (currently limited to parsing a maximum of 5 records).
}
\section{Single Hit}{

If a ChemiDPlus search results in a single match, the URL returned is the RN URL itself that is derived from the "h1" HTML node. For multiple matches, the "h1" HTML node returns blank.
}

\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[centipede]{strsplit}}
\code{\link[tibble]{as_tibble}},\code{\link[tibble]{tribble}}
\code{\link[tidyr]{extract}},\code{\link[tidyr]{pivot_wider}}
\code{\link[dplyr]{mutate}}
\code{\link[stringr]{str_remove}}

Other chemidplus parsing: 
\code{\link{is404}()},
\code{\link{isMultipleHits}()},
\code{\link{isNoRecord}()}
}
\concept{chemidplus parsing}
