% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSingleHit.R
\name{isSingleHit}
\alias{isSingleHit}
\title{#' Get the RN Number an URL for a Phrase}
\usage{
isSingleHit(response)
}
\description{
#' Get the RN Number an URL for a Phrase
}
\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
 \code{\link[centipede]{strsplit}}
 \code{\link[tibble]{as_tibble}},\code{\link[tibble]{tribble}}
 \code{\link[tidyr]{extract}},\code{\link[tidyr]{pivot_wider}}
 \code{\link[dplyr]{mutate}}
 \code{\link[stringr]{str_remove}}
}
