% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{load_ncit}
\alias{load_ncit}
\title{FUNCTION_TITLE}
\usage{
load_ncit(conn, file = "", append = TRUE)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
https://evs.nci.nih.gov/evs-download/thesaurus-downloads
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{send}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}

Other nci evs schema: 
\code{\link{load_all_mth}()}
}
\concept{nci evs schema}
