% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_drug_count.R
\name{log_drug_count}
\alias{log_drug_count}
\title{FUNCTION_TITLE}
\usage{
log_drug_count(conn)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{lsTables}},\code{\link[pg13]{readTable}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
\code{\link[skyscraper]{nci_count}}
\code{\link[tibble]{tibble}}

Other cancergov: 
\code{\link{get_dictionary_and_links}()},
\code{\link{get_drug_detail_links}()},
\code{\link{get_drug_dictionary}()},
\code{\link{get_drug_link_synonym}()},
\code{\link{get_drug_link_url}()},
\code{\link{process_drug_link_ncit}()},
\code{\link{process_drug_link_synonym}()},
\code{\link{process_drug_link_url}()}
}
\concept{cancergov}
