% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdp-internals.R
\name{log_registry_number}
\alias{log_registry_number}
\title{Log Registry Number Matches for a Search}
\usage{
log_registry_number(
  conn,
  raw_search_term,
  search_type = "contains",
  sleep_time = 3,
  schema = "chemidplus",
  verbose = TRUE
)
}
\arguments{
\item{conn}{Postgres connection object}

\item{raw_search_term}{Character string of length 1 to be searched in ChemiDPlus}

\item{sleep_time}{If the response argument is missing, the number seconds to pause after reading the URL, Default: 3}

\item{schema}{Schema that the returned data is written to, Default: 'chemidplus'}

\item{type}{Type of search available at \href{https://chem.nlm.nih.gov/chemidplus/}{ChemiDPlus}, Default: "contains"}
}
\value{
Each section is parsed by a respective skyscraper function that stores the scraped results in a table of the same name in a schema. If a connection argument is not provided, the results are returned as a dataframe in the R console.
}
\description{
All ChemiDPlus Scraping Functions operate on a Registry Number URL (\code{rn_url}). The initial search is logged to a "REGISTRY_NUMBER_LOG" Table. If the RN URL is then tested for 404 Status and logged to the "RN_URL_VALIDITY" Table. The major sections found at the ChemiDPlus site are: "Names and Synonyms", "Classification", "Registry Numbers", "Links to Resources" with these sections are written to their respective tables "NAMES_AND_SYNONYMS", "CLASSIFICATION", "REGISTRY_NUMBERS", and "LINKS_TO_RESOURCES".
}
\section{Registry Number Log Table}{

The REGISTRY_NUMBER_LOG Table is the landing table for any ChemiDPlus searches using skyscrape. It is the place where a source concept is searched based on a given set of parameters and all the possible Registry Numbers (RN) that source concept can be associated with in ChemiDPlus. The Registry Number then serves as a jump-off point from where a second RN URL is read and split based on the sections, and read into their corresponding ChemiDPlus Tables.

The Table logs the Raw Concept, the processed version of the Concept (ie removed spaces and error-throwing characters to generate a valid search URL for the Concept), the type of search (ie equals or contains), and the final search URL used to read a search result. A series of booleans are performed to determine whether the search was performed (ie a response was received), and if the results were for any records, and if these records were saved. If an RN was found, it is included along with the full URL associated with the URL.
}

\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{distinct}}
\code{\link[stringr]{str_remove}}
\code{\link[tibble]{tibble}}
\code{\link[xml2]{read_xml}}

Other chemidplus scraping: 
\code{\link{get_classification}()},
\code{\link{get_links_to_resources}()},
\code{\link{get_names_and_synonyms}()},
\code{\link{get_registry_numbers}()},
\code{\link{get_rn_url_validity}()}
}
\concept{chemidplus scraping}
