% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_functions.R
\name{scrape_pubmed}
\alias{scrape_pubmed}
\title{Scrape a PubMed search}
\usage{
scrape_pubmed(conn, search_term, max_return_size = 5)
}
\description{
The search is constrained to return results in the order of descending publication date.
}
\seealso{
\code{\link[stringr]{str_remove}}
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[tibble]{tibble}}
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{appendTable}},\code{\link[pg13]{writeTable}}
}
\concept{pubmed}
