% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemidplus_functions.R
\name{searchChemiDPlus}
\alias{searchChemiDPlus}
\title{Search ChemiDPlus and Store Results}
\usage{
searchChemiDPlus(
  conn,
  search_term,
  type = "contains",
  sleep_time = 5,
  schema = "chemidplus",
  export_repo = FALSE,
  target_dir = "~/GitHub/Public-Packages/chemidplusData/"
)
}
\description{
Search ChemiDPlus and Store Results
}
\seealso{
\code{\link[pg13]{query}},\code{\link[pg13]{buildQuery}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{filter}}
\code{\link[centipede]{no_na}}
\code{\link[tibble]{tribble}},\code{\link[tibble]{c("tibble", "tibble")}}
\code{\link[police]{try_catch_error_as_null}}
}
