% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyscraper-package.R
\docType{package}
\name{skyscraper-package}
\alias{skyscraper}
\alias{skyscraper-package}
\title{skyscraper: skyscraper R Package}
\description{
A package for scraping, storing, and querying drug data across disparate web and data sources
}
\section{Web Sources}{

ChemiDPlus: \href{https://chem.nlm.nih.gov/chemidplus/}{https://chem.nlm.nih.gov/chemidplus/}
NCI Drug Dictionary
UpToDate
}

\section{Data Sources}{

ClinicalTrials.gov
}

\keyword{internal}
